Add many changes from the katzenpost monorepo
Here we delete a bunch of stuff from the ref implementation such
that the go compiler is able to follow all the dependencies
and build correctly with the C components.

The alternative to this would be to build a .a or .so and install that somewhere
and then include that via CFLAGS so that the go compiler can build against it
and the relevant .h header files.

The motivation to do it this way instead is that it makes it possible
to install this via go install; if the requisite CGO_CFLAGS_ALLOW var is set.

+++ ref/randombytes.c
@@ -11,7 +11,7 @@ static int fd = -1;
 
 void randombytes(unsigned char *x, unsigned long long xlen)
 {
-    unsigned long long i;
+    int i;
 
     if (fd == -1) {
         for (;;) {
@@ -31,7 +31,7 @@ void randombytes(unsigned char *x, unsigned long long xlen)
             i = 1048576;
         }
 
-        i = (unsigned long long)read(fd, x, i);
+        i = read(fd, x, i);
         if (i < 1) {
             sleep(1);
             continue;
